requirements.txt
flask
flask-sqlalchemy
plotly
pandas
gunicorn
python-dotenv
reportlab